FIX: Use transformer for forcing small logo (#40)
Using a transformer is more reliable and less intrusive than using a plugin outlet to replace a chunk of a core template. The `home-logo-minimized` transformer is a recent addition to core (added in https://github.com/discourse/discourse/pull/30832), however this change is fully backward compatible -- if a site doesn't have the `registerValueTransformer` API or the new transformer, it will fallback to the plugin outlet override.